[category] (feat/delete-auth): Implemented role based authorization for DELETE /api/v1/categories/{category_id} api endpoint

+++ cortex/rest/handlers/delete_category_by_id.go
@@ -4,6 +4,7 @@ import (
 	"log/slog"
 	"net/http"
 
+	"cortex/rest/middlewares"
 	"cortex/rest/utils"
 
 	"github.com/google/uuid"
@@ -15,11 +16,18 @@ func (handlers *Handlers) DeleteCategoryByID(w http.ResponseWriter, r *http.Requ
 		utils.SendError(w, http.StatusBadRequest, "invalid UUID", nil)
 		return
 	}
+
+	if !middlewares.IsAdmin(r) {
+		utils.SendError(w, http.StatusForbidden, "Only admin can delete categories", nil)
+		return
+	}
+
 	if err := handlers.CategoryService.DeleteCategoryByUUID(r.Context(), category_uuid); err != nil {
-		slog.Error("handler: category deletation failed", slog.Any("error", err))
+		slog.Error("handler: category deletion failed", slog.Any("error", err))
 		utils.SendError(w, http.StatusInternalServerError, "failed to delete category", err)
 		return
 	}
+	
 	utils.SendJson(w, http.StatusOK, SuccessResponse{
 		Message: "Category deleted successfully",
 		Status:  http.StatusOK,
